requirements.txt
python-vlc >= 3.0.4106
Pillow
RPi.GPIO
Adafruit_GPIO

